update in resize

+++ Image/Image.py
@@ -504,7 +504,7 @@ class ImageTensor(Tensor):
         out = in_place_fct(self, in_place).reset_layers_order(in_place=False)
         if keep_ratio:
             ratio = torch.tensor(self.image_size) / torch.tensor(shape)
-            out.data = F.interpolate(out.to_tensor(), (shape[ratio.argmax()], shape[ratio.argmax()] / ratio.max()),
+            out.data = F.interpolate(out.to_tensor(), (shape[ratio.argmax()], int(shape[ratio.argmax()] / ratio.max())),
                                      mode='bilinear')  # , scale_factor=float(1 / ratio))
 
             out.pad([shape[0] - out.shape[-2], shape[1] - out.shape[-1]], in_place=True)
